+++ Tests/Tests/STPPaymentIntentTest.m
@@ -114,16 +114,16 @@
                    STPPaymentIntentConfirmationMethodUnknown);
 }
 
-- (void)testSourceActionFromString {
-    XCTAssertEqual([STPPaymentIntent sourceActionTypeFromString:@"authorize_with_url"],
-                   STPPaymentIntentSourceActionTypeAuthorizeWithURL);
-    XCTAssertEqual([STPPaymentIntent sourceActionTypeFromString:@"AUTHORIZE_WITH_URL"],
-                   STPPaymentIntentSourceActionTypeAuthorizeWithURL);
+- (void)testActionFromString {
+    XCTAssertEqual([STPPaymentIntent actionTypeFromString:@"redirect_to_url"],
+                   STPPaymentIntentActionTypeRedirectToURL);
+    XCTAssertEqual([STPPaymentIntent actionTypeFromString:@"REDIRECT_TO_URL"],
+                   STPPaymentIntentActionTypeRedirectToURL);
 
-    XCTAssertEqual([STPPaymentIntent confirmationMethodFromString:@"garbage"],
-                   STPPaymentIntentSourceActionTypeUnknown);
-    XCTAssertEqual([STPPaymentIntent confirmationMethodFromString:@"GARBAGE"],
-                   STPPaymentIntentSourceActionTypeUnknown);
+    XCTAssertEqual([STPPaymentIntent actionTypeFromString:@"garbage"],
+                   STPPaymentIntentActionTypeUnknown);
+    XCTAssertEqual([STPPaymentIntent actionTypeFromString:@"GARBAGE"],
+                   STPPaymentIntentActionTypeUnknown);
 }
 
 #pragma mark - Description Tests
@@ -178,15 +178,32 @@
     XCTAssertEqualObjects(paymentIntent.stripeDescription, @"My Sample PaymentIntent");
     XCTAssertFalse(paymentIntent.livemode);
     XCTAssertEqualObjects(paymentIntent.receiptEmail, @"danj@example.com");
+    
+    // Deprecated: nextSourceAction
+#pragma clang diagnostic push
+#pragma clang diagnostic ignored "-Wdeprecated"
     XCTAssertNotNil(paymentIntent.nextSourceAction);
     XCTAssertEqual(paymentIntent.nextSourceAction.type, STPPaymentIntentSourceActionTypeAuthorizeWithURL);
     XCTAssertNotNil(paymentIntent.nextSourceAction.authorizeWithURL);
     XCTAssertNotNil(paymentIntent.nextSourceAction.authorizeWithURL.url);
-    NSURL *returnURL = paymentIntent.nextSourceAction.authorizeWithURL.returnURL;
+    NSURL *nextSourceActionReturnURL = paymentIntent.nextSourceAction.authorizeWithURL.returnURL;
+    XCTAssertNotNil(nextSourceActionReturnURL);
+    XCTAssertEqualObjects(nextSourceActionReturnURL, [NSURL URLWithString:@"payments-example://stripe-redirect"]);
+    NSURL *nextSourceActionURL = paymentIntent.nextSourceAction.authorizeWithURL.url;
+    XCTAssertNotNil(nextSourceActionURL);
+#pragma clang diagnostic pop
+    
+    // nextAction
+    XCTAssertNotNil(paymentIntent.nextAction);
+    XCTAssertEqual(paymentIntent.nextAction.type, STPPaymentIntentActionTypeRedirectToURL);
+    XCTAssertNotNil(paymentIntent.nextAction.redirectToURL);
+    XCTAssertNotNil(paymentIntent.nextAction.redirectToURL.url);
+    NSURL *returnURL = paymentIntent.nextAction.redirectToURL.returnURL;
     XCTAssertNotNil(returnURL);
     XCTAssertEqualObjects(returnURL, [NSURL URLWithString:@"payments-example://stripe-redirect"]);
-    NSURL *url = paymentIntent.nextSourceAction.authorizeWithURL.url;
+    NSURL *url = paymentIntent.nextAction.redirectToURL.url;
     XCTAssertNotNil(url);
+
     XCTAssertEqualObjects(url, [NSURL URLWithString:@"https://hooks.stripe.com/redirect/authenticate/src_1Cl1AeIl4IdHmuTb1L7x083A?client_secret=src_client_secret_DBNwUe9qHteqJ8qQBwNWiigk"]);
     XCTAssertEqualObjects(paymentIntent.sourceId, @"src_1Cl1AdIl4IdHmuTbseiDWq6m");
     XCTAssertEqual(paymentIntent.status, STPPaymentIntentStatusRequiresAction);